# docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  models.Response:
    properties:
      message:
        type: string
    type: object
...
  utils.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
host: http://worklog-dev.odds.team/api
info:
  contact: {}
  description: This is a sample server odds-worklog server.
  license: {}
  title: Odds-Worklog Example API
  version: "1.0"
paths:
...
  /files/image:
    post:
      consumes:
      - image/png
      - image/gif
      - image/jpeg
      description: Upload image profile
      parameters:
      - description: file
        in: body
        name: image-profile
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Upload image profile
      tags:
      - files
  /files/image/{id}:
    delete:
      description: remove image file
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Remove image file
      tags:
      - files
    get:
      description: Download image file
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Download image file
      tags:
      - files
  /files/degreecertificate:
    post:
      consumes:
      - multipart/form-data
      description: Upload degree certificate file
      parameters:
      - description: file
        in: body
        name: file
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Upload transcript file
      tags:
      - files
  /files/degreecertificate/{id}:
    delete:
      description: remove degree certificate file
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Response'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Remove degree certificate file
      tags:
      - files
    get:
      description: Download degree certificate file
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Download degree certificate file
      tags:
      - files

      /files/idcard:
    post:
      consumes:
      - multipart/form-data
      description: Upload idcard file
      parameters:
      - description: file
        in: body
        name: file
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Upload idcard file
      tags:
      - files
  /files/idcard/{id}:
    delete:
      description: remove idcard file
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Response'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Remove idcard file
      tags:
      - files
    get:
      description: Download idcard file
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Download transcript file
      tags:
      - files
  /files/transcript:
    post:
      consumes:
      - multipart/form-data
      description: Upload transcript file
      parameters:
      - description: file
        in: body
        name: file
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Upload transcript file
      tags:
      - files
  /files/transcript/{id}:
    delete:
      description: remove transcript file
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Response'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Remove transcript file
      tags:
      - files
    get:
      description: Download transcript file
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Download transcript file
      tags:
      - files
...
swagger: "2.0"